+++ fessql/__init__.py
@@ -27,4 +27,4 @@ __all__ = (
     "__version__",
 )
 
-__version__ = "1.0.2b1"
+__version__ = "1.0.3b1"